+++ cli/cli.js
@@ -299,6 +299,8 @@ class DjinnJS {
                     data = data.replace('REPLACE_WITH_ENVIRONMENT', this.sites[i].env);
                     data = data.replace('REPLACE_WITH_GTAG_ID', this.sites[i].gtagId);
                     data = data.replace('REPLACE_WITH_DEFAULT_TRANSITION', this.sites[i].defaultTransition);
+                    data = data.replace("'REPLACE_WITH_PJAX_STATUS'", this.sites[i].disablePjax);
+                    data = data.replace("'REPLACE_WITH_PREFETCH_STATUS'", this.sites[i].disablePrefetching);
                     fs.writeFile(runtimeFile, data, error => {
                         if (error) {
                             reject(error);
@@ -439,6 +441,8 @@ class DjinnJS {
                     env: this.config.env,
                     gtagId: this.config.gtagId,
                     defaultTransition: this.config.defaultTransition,
+                    disablePjax: this.config.disablePjax,
+                    disablePrefetching: this.config.disablePrefetching,
                 };
                 configChecker(site)
                     .then(validSite => {